A few changes in order to fix the minor errors of the implementation of the DTO model

+++ src/main/java/com/store/store/dto/CustomerDto.java
@@ -1,23 +1,24 @@
 package com.store.store.dto;
 
+
 public class CustomerDto {
-    private Long id;
+    private int id;
     private String firstName;
     private String lastName;
     private String email;
 
-    public CustomerDto(Long id, String firstName, String lastName, String email) {
+    public CustomerDto(int id, String firstName, String lastName, String email) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
